refactor: NewCardScreenContent 컴포넌트 이름 변경 및 분리
